requirements.txt
pyyaml
boto3